# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WASAtext
  description: |-
    homework wasa
  version: 0.0.1
...
paths:
...
  /profiles/{user_id}/conversations/{dest_user_id}/messages:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/convo_id"
    get:
      summary: list all messages in a conversation
      tags: ["Conversation"]
      operationId: getMyMessages
      description: Obtain list of all messages in a conversation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: message list obtained succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/MessagesList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      parameters:
        - name: reply
          in: query
          required: false
          schema: {$ref: "#/components/schemas/Identificator"}
      tags: ["Conversation"]
      summary: send message
      operationId: sendMessage
      description: send new message in the selected conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Message"}
      responses:
        "200":
          description: message sent succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profiles/{user_id}/groups/{group_id}/messages:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/group_id"
    get:
      summary: list all messages in a group
      tags: ["Group"]
      operationId: getGroupMessages
      description: Obtain list of all messages in a group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: message list obtained succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/MessagesList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      parameters:
        - name: reply
          in: query
          required: false
          schema: {$ref: "#/components/schemas/Identificator"}
      tags: ["Group"]
      summary: send message
      operationId: sendGroupMessage
      description: send new message in the selected group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Message"}
      responses:
        "200":
          description: message sent succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  parameters: #tutto cio che prendi nei path
    user_id:
      name: user_id
      description: id of a given user
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identificator" }
    convo_id:
      name: dest_user_id
      description: id of a conversation user
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identificator" }
...
    group_id:
      name: group_id
      description: id of a given group
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identificator" }
  schemas:
    Identificator:
      description: id for user/conversation/group/message
      type: integer
      minimum: 1
      maximum: 999
      example: 9
...
    PhotoOrGif:
      description: image
      type: object
      properties:
        photo:
          description: image or gif
          type: string
          format: byte
          minLength: 0
          maxLength: 9999
...
    Message:
      description: user's message in a convo/group
      type: object
      properties:
        msg_id: { $ref: "#/components/schemas/Identificator" }
        timestamp: { $ref: "#/components/schemas/Timestamp" }
        content:
          oneOf:
            - $ref: "#/components/schemas/Text"
            - $ref: "#/components/schemas/PhotoOrGif"
      required: 
        - msg_id
        - timestamp
    MessagesList:
      type: object
      description: list of messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
          minItems: 0 
          maxItems: 999
      required:
        - messages 
...
    Text:
      type: object
      description: text of the message
      properties:
        text:
          type: string
          example: "Hey, come stai?"
          pattern: '^[^\/\\]{0,64}$'
          minLength: 1
          maxLength: 65536
...
    Timestamp:
      title: timestamp
      description: the timestamp of the message
      type: string
      format: date-time
      pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
      example: "2020-12-31T23:59:59"
      minLength: 19
      maxLength: 19
    Ok:
      title: OK
      description: The succesfull response from server.
      type: string
      example: "OK, action successfully completed"
      minLength: 2
      maxLength: 200
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
  responses:
...
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        text/plain:
          schema:
            description: BadRequest
            type: string
            example: "Bad Request"
            minLength: 11
            maxLength: 30

    #401
    Unauthorized:
      description: The access token is missing or it's expired.
      content:
        text/plain:
          schema:
            description: Unauthorized
            type: string
            example: "Unauthorized"
            minLength: 12
            maxLength: 30

    #403
    Forbidden:
      description: The user is not allowed to perform the action.
      content:
        text/plain:
          schema:
            description: Forbidden
            type: string
            example: "Forbidden"
            minLength: 9
            maxLength: 30

    #500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        text/plain:
          schema:
            description: InternalServerError
            type: string
            example: "Internal Server Error"
            minLength: 21
            maxLength: 30